chore(team): split remote relay grpc helpers
