+++ src/team/manager/remote_relay.rs
@@ -13,13 +13,10 @@ use serde_json::{Value, json};
 use sqlx::{Row, SqlitePool};
 
 use crate::agent::AGENT_NODE_MAIN_ID;
-use crate::internal::auth::{InternalAction, InternalAuthz, InternalAuthzConfig, InternalRole};
 use crate::internal::client::{
     InternalGrpcMailboxClient, InternalGrpcMailboxClientConfig, InternalGrpcPeerClientConfig,
 };
-use crate::internal::p2p::{
-    CredentialProvider, NodeCredentialRequest, P2PTransport, build_message_metadata,
-};
+use crate::internal::p2p::{P2PTransport, build_message_metadata};
 use crate::internal::tls::InternalGrpcSecurityMode;
 use crate::team::TeamActorMessageRecord;
 
@@ -34,11 +31,12 @@ pub(super) const RELAY_TIMEOUT_MAX_MS: u64 = 60_000;
 
 #[derive(Clone)]
 pub(super) struct TeamRemoteRelayAdapter {
-    db: SqlitePool,
-    http_client: reqwest::Client,
-    grpc_tls_defaults: Arc<Mutex<Option<GrpcRelayTlsDefaults>>>,
-    grpc_peer_client_config: Arc<Mutex<Option<InternalGrpcPeerClientConfig>>>,
-    grpc_client_cache: Arc<Mutex<HashMap<GrpcRelayClientCacheKey, InternalGrpcMailboxClient>>>,
+    pub(super) db: SqlitePool,
+    pub(super) http_client: reqwest::Client,
+    pub(super) grpc_tls_defaults: Arc<Mutex<Option<GrpcRelayTlsDefaults>>>,
+    pub(super) grpc_peer_client_config: Arc<Mutex<Option<InternalGrpcPeerClientConfig>>>,
+    pub(super)
+        grpc_client_cache: Arc<Mutex<HashMap<GrpcRelayClientCacheKey, InternalGrpcMailboxClient>>>,
 }
 
 #[derive(Debug, Clone, PartialEq, Eq)]
@@ -306,170 +304,16 @@ impl TeamRemoteRelayAdapter {
         Ok(())
     }
 
-    async fn resolve_registered_grpc_route(
-        &self,
-        target_node_id: &str,
-        route: &GrpcRemoteRelayRouteValue,
-    ) -> Result<ResolvedGrpcRelayRoute, ActorRelayError<TeamRemoteRelayError>> {
-        let normalized_target_node_id = target_node_id.trim();
-        if normalized_target_node_id.is_empty() || normalized_target_node_id == AGENT_NODE_MAIN_ID {
-            return Err(ActorRelayError::permanent(
-                TeamRemoteRelayError::InvalidRoute(
-                    "route target_node_id must reference a registered remote agent node"
-                        .to_string(),
-                ),
-            ));
-        }
-
-        let row = sqlx::query(
-            r#"
-            SELECT grpc_target, tls_server_name
-            FROM agent_nodes
-            WHERE id = ?1
-            "#,
-        )
-        .bind(normalized_target_node_id)
-        .fetch_optional(&self.db)
-        .await
-        .map_err(|err| {
-            ActorRelayError::permanent(TeamRemoteRelayError::InvalidRoute(format!(
-                "resolve registered gRPC route failed: {err}"
-            )))
-        })?
-        .ok_or_else(|| {
-            ActorRelayError::permanent(TeamRemoteRelayError::InvalidRoute(format!(
-                "target agent node '{}' is not registered for gRPC relay",
-                normalized_target_node_id
-            )))
-        })?;
-
-        let registered_grpc_target = row.get::<String, _>("grpc_target");
-        let registered_tls_server_name = row
-            .try_get::<Option<String>, _>("tls_server_name")
-            .ok()
-            .flatten()
-            .map(|value| value.trim().to_string())
-            .filter(|value| !value.is_empty());
-        let requested_grpc_target = route.grpc_target.trim();
-        if requested_grpc_target.is_empty() {
-            return Err(ActorRelayError::permanent(
-                TeamRemoteRelayError::MissingGrpcTarget,
-            ));
-        }
-        let url = Url::parse(requested_grpc_target)
-            .map_err(|_| ActorRelayError::permanent(TeamRemoteRelayError::InvalidGrpcTarget))?;
-        if url.scheme() != "https" {
-            return Err(ActorRelayError::permanent(
-                TeamRemoteRelayError::InvalidGrpcTarget,
-            ));
-        }
-        if requested_grpc_target != registered_grpc_target {
-            return Err(ActorRelayError::permanent(
-                TeamRemoteRelayError::InvalidRoute(format!(
-                    "route.grpc_target must match registered agent node '{}'",
-                    normalized_target_node_id
-                )),
-            ));
-        }
-
-        let requested_tls_server_name = route
-            .tls_server_name
-            .as_deref()
-            .map(str::trim)
-            .filter(|value| !value.is_empty());
-        if requested_tls_server_name != registered_tls_server_name.as_deref() {
-            return Err(ActorRelayError::permanent(
-                TeamRemoteRelayError::InvalidRoute(format!(
-                    "route.tls_server_name must match registered agent node '{}'",
-                    normalized_target_node_id
-                )),
-            ));
-        }
-
-        Ok(ResolvedGrpcRelayRoute {
-            grpc_target: registered_grpc_target,
-            tls_server_name: registered_tls_server_name,
-        })
-    }
-
-    fn issue_registered_grpc_access_token(
-        &self,
-    ) -> Result<String, ActorRelayError<TeamRemoteRelayError>> {
-        let peer_config = self
-            .grpc_peer_client_config
-            .lock()
-            .expect("lock grpc peer client config")
-            .clone()
-            .ok_or_else(|| {
-                ActorRelayError::permanent(TeamRemoteRelayError::GrpcPeerClientUnavailable)
-            })?;
-        let authz = InternalAuthz::new(InternalAuthzConfig {
-            shared_secret: peer_config.shared_secret.clone(),
-            expected_issuer: peer_config.expected_issuer.clone(),
-            expected_audience: peer_config.expected_audience.clone(),
-        });
-        let issued = authz
-            .issue_node_access_token(NodeCredentialRequest {
-                source_node_id: peer_config.source_node_id.clone(),
-                role: InternalRole::Coordinator.as_str().to_string(),
-                actor_id: None,
-                run_id: None,
-                permissions: vec![InternalAction::MessageSend.as_str().to_string()],
-                scope: Vec::new(),
-                audience: Vec::new(),
-                ttl_seconds: 600,
-            })
-            .map_err(|err| {
-                ActorRelayError::permanent(TeamRemoteRelayError::InvalidRoute(format!(
-                    "issue relay access token failed: {err}"
-                )))
-            })?;
-        Ok(issued.access_token)
-    }
-
-    async fn grpc_client_for_route(
-        &self,
-        config: InternalGrpcMailboxClientConfig,
-    ) -> Result<InternalGrpcMailboxClient, ActorRelayError<TeamRemoteRelayError>> {
-        let cache_key = GrpcRelayClientCacheKey {
-            target: config.target.clone(),
-            access_token: config.access_token.clone(),
-            ca_cert_path: config.ca_cert_path.clone(),
-            tls_server_name: config.tls_server_name.clone(),
-            client_cert_path: config.client_cert_path.clone(),
-            client_key_path: config.client_key_path.clone(),
-        };
-        if let Some(client) = self
-            .grpc_client_cache
-            .lock()
-            .expect("lock grpc client cache")
-            .get(&cache_key)
-            .cloned()
-        {
-            return Ok(client);
-        }
-
-        let client = InternalGrpcMailboxClient::connect(config)
-            .await
-            .map_err(|err| {
-                ActorRelayError::retryable(TeamRemoteRelayError::GrpcConnect(err.to_string()))
-            })?;
-        self.grpc_client_cache
-            .lock()
-            .expect("lock grpc client cache")
-            .insert(cache_key, client.clone());
-        Ok(client)
-    }
 }
 
 #[derive(Debug, Clone, PartialEq, Eq, Hash)]
-struct GrpcRelayClientCacheKey {
-    target: String,
-    access_token: String,
-    ca_cert_path: Option<String>,
-    tls_server_name: Option<String>,
-    client_cert_path: Option<String>,
-    client_key_path: Option<String>,
+pub(super) struct GrpcRelayClientCacheKey {
+    pub(super) target: String,
+    pub(super) access_token: String,
+    pub(super) ca_cert_path: Option<String>,
+    pub(super) tls_server_name: Option<String>,
+    pub(super) client_cert_path: Option<String>,
+    pub(super) client_key_path: Option<String>,
 }
 
 #[derive(Debug, thiserror::Error)]
@@ -557,12 +401,6 @@ pub(super) struct GrpcRemoteRelayRouteValue {
     pub(super) access_token: Option<String>,
     #[serde(default)]
     pub(super) tls_server_name: Option<String>,
-}
-
-#[derive(Debug)]
-struct ResolvedGrpcRelayRoute {
-    grpc_target: String,
-    tls_server_name: Option<String>,
 }
 
 #[derive(Debug, Deserialize)]
